refactor(web): lift HeroUIProvider + LoginModal to locale root, merge 401 handling
Two simplifications now that the auth-UI architecture has settled:

1. HeroUIProvider and LoginModal both move to a new RootProviders client
   wrapper rendered from [locale]/layout.tsx. They used to be mounted in
   each of LandingProvidersLayout and ProvidersLayout — duplicated config
   that only ever ran one branch at a time. One mount, available to every
   route group, still lazy-loaded so it stays out of the initial bundle.
   The dev/layout no longer needs to re-wrap HeroUIProvider either.

2. `useUnauthorizedChallenge` is folded back into `processAxiosError`.
   It was a separate hook only when the two halves of error handling had
   different mount conditions; now both are (main)-only, so two response
   interceptors on the same axios instance with identical lifecycle is
   pure overhead. One hook, one interceptor, one switch statement.

Entire-Checkpoint: 5b4d780bdec1
